created all members subcomponent to becalled on map, edited make post styles
