+++ src/styles/allmembersactivity.css
@@ -151,3 +151,42 @@
     font-size: 1px;
     margin-bottom: 1vh;
 }
+
+.comment{
+    display: flex;
+}
+.textAreacomment{
+    width: 90%;
+    margin-left: .5vw;
+    border-radius: 15px;
+    border: 1px solid  #9e9b9b;
+    padding: 0px 14px;
+    padding-top: 1px;
+    font-family: sans-serif;
+    font-size: 13px;
+    color: #726e6e;
+}
+.textAreacomment:focus{
+    outline-color: #8931E5;
+    outline: none;
+    border: 1px solid #8931E5;
+}
+
+.btnsPost{
+    width:100%;
+}
+
+.btnbtnbtn{
+    display: flex;
+    margin-left: 5px;
+    margin-top: 1vh;
+    align-items: center;
+}
+.btnbtnbtn button{
+    font-size: 13px;
+}
+.submitPost-comment{
+    padding: 5px 20px;
+    margin-right: 1vw;
+    /* box-shadow: 10px 10px 2px rgba(0, 0, 0px, 0.5); */
+}
